complted ats resume-scan

+++ src/main/java/careerpilot_parent/ats/dto/response/AtsScanResponse.java
@@ -1,10 +1,12 @@
 package careerpilot_parent.ats.dto.response;
 
-
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.time.LocalDateTime;
-
 
 @Getter
 @Setter
@@ -13,32 +15,21 @@ import java.time.LocalDateTime;
 @Builder
 public class AtsScanResponse {
 
-
     private Long id;
-
 
     private Long resumeId;
 
-
     private String jobTitle;
-
 
     private String companyName;
 
-
-    private Integer atsScore;
-
+    private Double atsScore;
 
     private String matchedSkills;
 
-
     private String missingSkills;
-
 
     private String suggestions;
 
-
     private LocalDateTime createdAt;
-
-
 }
